Процессор теперь может работать с частотй не привязаной к rmii
Частота процессора поднята до 70 MHz
Добавлена возможность выбора типа значения частоты float/double
В модуле arp добавлены ifdef'ы для PUCKED_STRUCT
Изменена обработка прерывания частотомера с целью уменьшить количество
входов в него, но удлиннить один цикл обработки

+++ application/libhw/settings.h
@@ -30,7 +30,7 @@ struct sSettings {
     uint8_t MAC_ADDR[MAC_ADDRESS_SIZE];
     uint8_t DHCP;
 
-    double ReferenceFrequency;
+    uint32_t ReferenceFrequency;
 
     uint32_t CRC32; // 4 bytes
 };
